requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.28.0
bs4>=0.0.1
lxml>=4.9.0
pytz>=2023.3

# API and Web Framework
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
python-multipart>=0.0.6
websockets>=11.0.0

# ML & DL Libraries
torch>=2.0.0
transformers>=4.28.0
xgboost>=1.7.0
shap>=0.41.0
stable-baselines3>=1.8.0
ta>=0.10.0  # Technical Analysis library
yfinance>=0.2.18
pyotp>=2.8.0  # For Angel One API authentication

# For news scraping and sentiment analysis
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
sentencepiece>=0.1.99
finbert>=0.0.5

# Testing
pytest>=7.3.0

# Deployment
gunicorn>=20.1.0